refactor: H.264 sample entry生成処理を共通化
