+++ src/encoder_openh264.rs
@@ -1,18 +1,11 @@
 use std::sync::Arc;
 
 use orfail::OrFail;
-use shiguredo_mp4::{
-    Uint,
-    boxes::{Avc1Box, AvccBox, SampleEntry},
-};
 
 use crate::{
     encoder::VideoEncoderOptions,
-    video::{self, VideoFormat, VideoFrame},
-    video_h264::{
-        H264_LEVEL_3_1, H264_NALU_TYPE_PPS, H264_NALU_TYPE_SEI, H264_NALU_TYPE_SPS,
-        H264_PROFILE_BASELINE, H264AnnexBNalUnits, NALU_HEADER_LENGTH,
-    },
+    video::{VideoFormat, VideoFrame},
+    video_h264::{self, H264_NALU_TYPE_SEI, H264AnnexBNalUnits},
 };
 
 #[derive(Debug)]
@@ -56,7 +49,10 @@ impl Openh264Encoder {
 
         let sample_entry = if self.is_first {
             self.is_first = false;
-            Some(sample_entry(frame.width, frame.height, &encoded.data).or_fail()?)
+            Some(
+                video_h264::h264_sample_entry_from_annexb(frame.width, frame.height, &encoded.data)
+                    .or_fail()?,
+            )
         } else {
             None
         };
@@ -97,40 +93,4 @@ impl Openh264Encoder {
     pub fn next_encoded_frame(&mut self) -> Option<VideoFrame> {
         self.encoded.take()
     }
-}
-
-fn sample_entry(width: usize, height: usize, data: &[u8]) -> orfail::Result<SampleEntry> {
-    // H.264 ストリームから SPS と PPS と取り出す
-    let mut sps_list = Vec::new();
-    let mut pps_list = Vec::new();
-    for nalu in H264AnnexBNalUnits::new(data) {
-        let nalu = nalu.or_fail()?;
-        match nalu.ty {
-            H264_NALU_TYPE_SPS => sps_list.push(nalu.data.to_vec()),
-            H264_NALU_TYPE_PPS => pps_list.push(nalu.data.to_vec()),
-            _ => {}
-        }
-    }
-    (!sps_list.is_empty()).or_fail()?;
-    (!pps_list.is_empty()).or_fail()?;
-
-    Ok(SampleEntry::Avc1(Avc1Box {
-        visual: video::sample_entry_visual_fields(width, height),
-        avcc_box: AvccBox {
-            // 実際のエンコードストリームに合わせた値
-            sps_list,
-            pps_list,
-
-            // 以下は Hisui では固定値
-            avc_profile_indication: H264_PROFILE_BASELINE,
-            avc_level_indication: H264_LEVEL_3_1,
-            profile_compatibility: 0, // いったん 0 を指定しているが、もし支障があれば調整する
-            length_size_minus_one: Uint::new(NALU_HEADER_LENGTH as u8 - 1),
-            chroma_format: None,
-            bit_depth_luma_minus8: None,
-            bit_depth_chroma_minus8: None,
-            sps_ext_list: Vec::new(),
-        },
-        unknown_boxes: Vec::new(),
-    }))
 }
